🛠️FIX(layout): 사이드바로 api명세 전환 시 js 미동작 #7
기술부채 처리, 사이드바 클릭시 리다이렉트로 수정조치
